Add :db_dump_extra_options variable
At @stackbuilders we use this library to sync production db with
staging on many projects. On one of them we wanted to skip one (big)
table in the db dump. With this patch you can do it like this:

    $ DB_DUMP_EXTRA_OPTIONS="--exclude-table=logs" cap production db:pull

This will give flexibility to users without having to patch this gem
(e.g. clean flag in https://github.com/sgruhier/capistrano-db-tasks/pull/29)